aws_cdk_3tier_stack/requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_autoscaling
aws_cdk.aws_s3
aws_cdk.aws_ssm
aws_cdk.aws_rds